refactor: Update error handling and improve code structure
- Modified error handling in `main.go` to ensure proper logging and exit on error.
- Removed unnecessary build tag in `wire.go` files for cleaner code.
- Updated command registration in `cli/main.go` for consistency.
- Refactored agent service implementation to use `AgentServiceImpl` for better clarity and structure.
- Enhanced configuration structures in `config.go` files for improved readability and consistency.
- Introduced context support in repository interfaces for better handling of timeouts and cancellations.
- Added new `GatewayServiceImpl` for handling event streaming to the gateway.
- Updated various repository methods to follow context-based patterns for improved performance and maintainability.

+++ pkg/nova/delay_kafka.go
@@ -76,7 +76,7 @@ func NewDelayTopicManager(
 		codec:        DefaultMessageCodec, // Default to JSON codec
 		ctx:          ctx,
 		cancel:       cancel,
-		timerWheel:   NewTimerWheel(DefaultDelaySlotCount, int64(DefaultDelaySlotDuration.Milliseconds())),
+		timerWheel:   NewTimerWheel(DefaultDelaySlotCount, DefaultDelaySlotDuration.Milliseconds()),
 	}
 
 	// Generate delay topic names
